udatepost 수정, 세션 검사 기능 추가

+++ src/main/java/onboarding/crud/comment/controller/CommentController.java
@@ -1,10 +1,13 @@
 package onboarding.crud.comment.controller;
 
+import jakarta.servlet.http.HttpServletRequest;
 import onboarding.crud.comment.dto.CommentDto;
 import onboarding.crud.comment.service.CommentService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -20,13 +23,18 @@ public class CommentController {
         return commentService.getCommentsByPostId(postId);
     }
     @PostMapping
-    public CommentDto createComment(@PathVariable Long postId, @RequestBody CommentDto commentDto) {
-        //todo:initialize field
+    public CommentDto createComment(@PathVariable Long postId, @RequestBody CommentDto commentDto, HttpServletRequest request) {
+        Object _id = request.getSession().getAttribute("userId");
+        if(_id == null) throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+
         return commentService.createComment(postId, commentDto);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteComment(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteComment(@PathVariable Long id, HttpServletRequest request) {
+        Object _id = request.getSession().getAttribute("userId");
+        if(_id == null) throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+
         commentService.deleteComment(id);
         return ResponseEntity.noContent().build();
     }
